Base no testeada

+++ src/controllers/product.controller.js
@@ -1,4 +1,4 @@
-import { ProductService } from "../repositories/product.repository.js";
+import ProductService from "../repositories/product.repository.js";
 
 export const get = async(req, res) => {
     const products = await ProductService.get()
@@ -6,13 +6,15 @@ export const get = async(req, res) => {
 }
 
 export const create = async(req, res) => {
-    const product = req.body
+    const { name, brand, price, stock, category, thumbnail, description } = req.body
+    if (!name || !brand || !price || !stock || !category) return res.status(400).send({ status: 'error', error: 'Missing params' })
     const productNew = await ProductService.create(product)
     res.json({ product: productNew })
 }
 
 export const getById = async(req, res) => {
     const { id } = req.params
+    if (!id) return res.status(400).send({ status: 'error', error: 'Missing params' })
     const product = await ProductService.getById(id)
     res.json({ product })
 }
@@ -20,12 +22,14 @@ export const getById = async(req, res) => {
 export const updata = async(req, res) => {
     const { id } = req.params
     const product = req.body
+    if (!id) return res.status(400).send({ status: 'error', error: 'Missing params' })
     const productUpdata = await ProductService.updata(id, product)
     res.json({ product: productUpdata })
 }
 
 export const deleteById = async(req, res) => {
     const { id } = req.params
+    if (!id) return res.status(400).send({ status: 'error', error: 'Missing params' })
     const product = await ProductService.deleteById(id)
     res.json({ product })
 }
